telegram-admin-bot/requirements.txt
Flask==3.0.0
python-telegram-bot==20.7
python-dotenv==1.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3